feat(xo6): rename file for conresponding to host dashboard
