+++ @xen-orchestra/web/src/components/pool/dashboard/PoolDashboardpatch.vue
