+++ @xen-orchestra/web/src/components/pool/dashboard/PoolDashboardstatus.vue
